requirements.txt
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.2.1
openpyxl==3.1.2
jinja2==3.1.3
slack_sdk==3.27.1
python-dotenv==1.0.1
